// src/app/interfaces/interfaces.ts
export interface userDatasFirstComponentProps {
    name?: string
    age?: number
    phone?: number
    email?: string
    isFirstFormsComplete?: boolean
  }

export interface userDatasSecondComponentProps {
  city?: string
  district?: string
  street?: string
  numberOfTheHouse?: number
  isSecondFormsComplete?: boolean
}